requirements_ROS.txt
torch==1.0.1.post2
torchvision==0.2.1
tensorlayer==1.11.0
tensorflow-gpu==1.14.0
scipy==1.1.0
Pillow==6.2.0
numpy==1.15.4
